Use SSR component to retrieve Google Sheets data
