+++ views/ReadingLogView/ReadingLogView.tsx
@@ -1,0 +1,3 @@
+export default function ReadingLogView({ readingLogEntries }: any) {
+  return <>{readingLogEntries}</>;
+}
